Refactor pose and frame resolution helpers.
This adds `GraphResolver` which creates a more convenient interface for
resolving poses and frames in a way that can be overridden by tests. A
global instance `graph_resolver` is used by all functions so that we
don't have to pass `pose_resolver`, `axis_resolver` and `frame_resolver`
(in upcoming PRs) to all the `add_*` converter functions.

Signed-off-by: Addisu Z. Taddese <addisu@openrobotics.org>