Add name prefixes to geometries and freejoints
It is possible for SDFormat files to have collisions or visuals in
different links have the same name, but this is not allowed in MJCF. To
prevent this name collision, we prefix the names of geoms with the names
of their parent bodies.

This also changes how freejoints are named since it is possible to have
multiple freejoints in the world.

The delimiter `_` is chosen as opposed to `/` or `::` because these two
have special meaning in MJCF and SDFormat respectively.

Signed-off-by: Addisu Z. Taddese <addisu@openrobotics.org>