SDFormat to MJCF: Add support for force/torque sensors (#50)
The handling of //force_torque/frame is a little tricky. My interpretation of the SDFormat spec is that the reported torque is always about the joint origin, so I'm setting the pos of the mjcf site for the sensor to the joint origin regardless of the value in //force_torque/frame. For the rotation, I'm setting the rotation of the mjcf site for the sensor to the rotation of the body indicated by //force_torque/frame.
Signed-off-by: Addisu Z. Taddese <addisu@openrobotics.org>