Restructure package (Alternative to #39) (#74)
The proposed package layout looks like:

└── sdformat_mjcf
    ├── __init__.py
    ├── mjcf_to_sdformat
    │   └── __init__.py
    ├── sdformat_to_mjcf
    │   └── __init__.py
    └── utils

Where we have sdformat_mjcf as the top level package. For the directory structure for for the distribution package (gz-mujoco/sdformat_mjcf), I opted to use the src layout as encouraged in the pytest good practices page and explained in this blog. The advantage of the src layout is that our tests always test what's installed, not what's in the src directory.

In addition to shuffling files and fixing imports, I made a change to how TEST_RESOURCE_DIR is computed. This was necessary because the way we had it before, __file__ was being evaluated in the context of the importing module and so, was dependent on the location of the importing module. The intended behavior was that __file__ would always evaluate to the location of the helper.py module.

Signed-off-by: Addisu Z. Taddese <addisu@openrobotics.org>