Handle default geom types and mesh scales (#95)
While trying to convert dm_control/dm_control/suite/dog.xml I noticed that we are not handling geoms without a type attribute. The default value of geom types is sphere according to the MJCF docs.

Apparently it's also possible to specify geom[@mesh] without setting type, and this PR also addresses that.

Finally, we were not checking if geom[@scale] was set when processing meshes. This PR applies the default scale if the attribute is not set.

Signed-off-by: Addisu Z. Taddese <addisu@openrobotics.org>